+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed Players are not always teleported to lobby location after falling into void
 * Fixed NPE (Cannot invoke) if no cage on base
 * Fixed fireticks are not removed after jumped into lava portal
+* Fixed arena does not end if there are only players in one base left
 
 ### 1.1.1 Release (23.04.2021 - 10.05.2021)
 * Added /tb join maxplayers which tries to get the arena with the highest amount of players